Make sure Traefik gets triggered if Docker recovers after a failure.
Before, if Docker failed to start, Traefik would fail too since we list Docker
as a dependency. A unit failing like this with status "dependency", does not
trigger its own restarting mechanism (it is deemed unrecoverable?).
With this change, we explicitly trigger Traefik whenever Docker is started,
such that if Docker recovers after a failure, Traefik will get triggered too.